Implemented feature to BoW representation with pooling
The process of assigning a feature vector to
codebook. This is done by checking for each row
in the feature vector to which cluster index it belongs.
Then we compute the distance to the related cluster and
do max or sum pooling.

Future patch should implement other representation
and more pooling methods.